Refactored verifier
