+++ tlsn/tlsn-verifier/src/tls/mod.rs
@@ -2,23 +2,19 @@
 
 pub(crate) mod config;
 mod error;
+mod future;
+mod notarize;
 pub mod state;
+mod verify;
 
 pub use config::{VerifierConfig, VerifierConfigBuilder, VerifierConfigBuilderError};
 pub use error::VerifierError;
 
-use std::{
-    pin::Pin,
-    time::{SystemTime, UNIX_EPOCH},
-};
-
+use crate::{tls::future::OTFuture, Mux};
 use futures::{
-    future::FusedFuture,
     stream::{SplitSink, SplitStream},
-    AsyncRead, AsyncWrite, Future, FutureExt, SinkExt, StreamExt, TryFutureExt,
+    AsyncRead, AsyncWrite, FutureExt, StreamExt, TryFutureExt,
 };
-
-use mpz_core::serialize::CanonicalSerialize;
 use mpz_garble::{config::Role as GarbleRole, protocol::deap::DEAPVm};
 use mpz_ot::{
     actor::kos::{
@@ -27,21 +23,18 @@ use mpz_ot::{
     chou_orlandi, kos,
 };
 use mpz_share_conversion as ff;
-use mpz_share_conversion::ShareConversionVerify;
 use rand::Rng;
 use signature::Signer;
+use std::time::{SystemTime, UNIX_EPOCH};
 use tls_mpc::{setup_components, MpcTlsFollower, TlsRole};
-use tlsn_core::{
-    msg::{SignedSessionHeader, TlsnMessage},
-    HandshakeSummary, SessionHeader, Signature,
-};
+use tlsn_core::{SessionHeader, Signature};
 use uid_mux::{yamux, UidYamux};
-use utils_aio::{codec::BincodeMux, duplex::Duplex, expect_msg_or_err, mux::MuxChannel};
-
-use crate::Mux;
+use utils_aio::{codec::BincodeMux, duplex::Duplex, mux::MuxChannel};
 
 #[cfg(feature = "tracing")]
 use tracing::{debug, info, instrument};
+
+use self::future::MuxFuture;
 
 type OTSenderActor = SenderActor<
     chou_orlandi::Receiver,
@@ -123,7 +116,8 @@ impl Verifier<state::Initialized> {
             .await?
             .run()
             .await?
-            .notarize(signer)
+            .start_notarize()
+            .finalize(signer)
             .await
     }
 }
@@ -186,92 +180,15 @@ impl Verifier<state::Setup> {
 }
 
 impl Verifier<state::Closed> {
-    /// Notarizes the TLS session.
-    pub async fn notarize<T>(self, signer: &impl Signer<T>) -> Result<SessionHeader, VerifierError>
-    where
-        T: Into<Signature>,
-    {
-        let state::Closed {
-            mut mux,
-            mut mux_fut,
-            mut vm,
-            ot_send,
-            ot_recv,
-            ot_fut,
-            mut gf2,
-            encoder_seed,
-            start_time,
-            server_ephemeral_key,
-            handshake_commitment,
-            sent_len,
-            recv_len,
-        } = self.state;
-
-        let notarize_fut = async {
-            let mut notarize_channel = mux.get_channel("notarize").await?;
-
-            let merkle_root =
-                expect_msg_or_err!(notarize_channel, TlsnMessage::TranscriptCommitmentRoot)?;
-
-            // Finalize all MPC before signing the session header
-            let (mut ot_sender_actor, _, _) = futures::try_join!(
-                ot_fut,
-                ot_send.shutdown().map_err(VerifierError::from),
-                ot_recv.shutdown().map_err(VerifierError::from)
-            )?;
-
-            ot_sender_actor.reveal().await?;
-
-            vm.finalize()
-                .await
-                .map_err(|e| VerifierError::MpcError(Box::new(e)))?;
-
-            gf2.verify()
-                .await
-                .map_err(|e| VerifierError::MpcError(Box::new(e)))?;
-
-            #[cfg(feature = "tracing")]
-            info!("Finalized all MPC");
-
-            let handshake_summary =
-                HandshakeSummary::new(start_time, server_ephemeral_key, handshake_commitment);
-
-            let session_header = SessionHeader::new(
-                encoder_seed,
-                merkle_root,
-                sent_len,
-                recv_len,
-                handshake_summary,
-            );
-
-            let signature = signer.sign(&session_header.to_bytes());
-
-            #[cfg(feature = "tracing")]
-            info!("Signed session header");
-
-            notarize_channel
-                .send(TlsnMessage::SignedSessionHeader(SignedSessionHeader {
-                    header: session_header.clone(),
-                    signature: signature.into(),
-                }))
-                .await?;
-
-            #[cfg(feature = "tracing")]
-            info!("Sent session header");
-
-            Ok::<_, VerifierError>(session_header)
-        };
-
-        let session_header = futures::select! {
-            res = notarize_fut.fuse() => res?,
-            _ = &mut mux_fut => Err(std::io::Error::from(std::io::ErrorKind::UnexpectedEof))?,
-        };
-
-        let mut mux = mux.into_inner();
-
-        futures::try_join!(mux.close().map_err(VerifierError::from), mux_fut)?;
-
-        Ok(session_header)
+    /// Starts notarization of the TLS session.
+    ///
+    /// If the verifier is a Notary, this function will transition the verifier to the next state
+    /// where it can sign the prover's commitments to the transcript.
+    pub fn start_notarize(self) -> Verifier<state::Notarize> {
+        Verifier {
+            config: self.config,
+            state: self.state.into(),
+        }
     }
 }
 
@@ -394,48 +311,4 @@ async fn setup_mpc_backend(
     debug!("MPC backend setup complete");
 
     Ok((mpc_tls, vm, ot_send, ot_recv, gf2, ot_fut))
-}
-
-/// A future which must be polled for the muxer to make progress.
-pub(crate) struct MuxFuture {
-    fut: Pin<Box<dyn FusedFuture<Output = Result<(), VerifierError>> + Send + 'static>>,
-}
-
-impl Future for MuxFuture {
-    type Output = Result<(), VerifierError>;
-
-    fn poll(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Self::Output> {
-        self.fut.as_mut().poll(cx)
-    }
-}
-
-impl FusedFuture for MuxFuture {
-    fn is_terminated(&self) -> bool {
-        self.fut.is_terminated()
-    }
-}
-
-/// A future which must be polled for the Oblivious Transfer protocol to make progress.
-pub(crate) struct OTFuture {
-    fut: Pin<Box<dyn FusedFuture<Output = Result<OTSenderActor, VerifierError>> + Send + 'static>>,
-}
-
-impl Future for OTFuture {
-    type Output = Result<OTSenderActor, VerifierError>;
-
-    fn poll(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Self::Output> {
-        self.fut.as_mut().poll(cx)
-    }
-}
-
-impl FusedFuture for OTFuture {
-    fn is_terminated(&self) -> bool {
-        self.fut.is_terminated()
-    }
 }
