+++ tlsn/tlsn-verifier/src/tls/notarize.rs
@@ -1,0 +1,103 @@
+use super::{state::Notarize, Verifier, VerifierError};
+use futures::{FutureExt, SinkExt, StreamExt, TryFutureExt};
+use mpz_core::serialize::CanonicalSerialize;
+use mpz_share_conversion::ShareConversionVerify;
+use signature::Signer;
+use tlsn_core::{
+    msg::{SignedSessionHeader, TlsnMessage},
+    HandshakeSummary, SessionHeader, Signature,
+};
+use utils_aio::{expect_msg_or_err, mux::MuxChannel};
+
+#[cfg(feature = "tracing")]
+use tracing::info;
+
+impl Verifier<Notarize> {
+    /// Notarizes the TLS session.
+    pub async fn finalize<T>(self, signer: &impl Signer<T>) -> Result<SessionHeader, VerifierError>
+    where
+        T: Into<Signature>,
+    {
+        let Notarize {
+            mut mux,
+            mut mux_fut,
+            mut vm,
+            ot_send,
+            ot_recv,
+            ot_fut,
+            mut gf2,
+            encoder_seed,
+            start_time,
+            server_ephemeral_key,
+            handshake_commitment,
+            sent_len,
+            recv_len,
+        } = self.state;
+
+        let notarize_fut = async {
+            let mut notarize_channel = mux.get_channel("notarize").await?;
+
+            let merkle_root =
+                expect_msg_or_err!(notarize_channel, TlsnMessage::TranscriptCommitmentRoot)?;
+
+            // Finalize all MPC before signing the session header
+            let (mut ot_sender_actor, _, _) = futures::try_join!(
+                ot_fut,
+                ot_send.shutdown().map_err(VerifierError::from),
+                ot_recv.shutdown().map_err(VerifierError::from)
+            )?;
+
+            ot_sender_actor.reveal().await?;
+
+            vm.finalize()
+                .await
+                .map_err(|e| VerifierError::MpcError(Box::new(e)))?;
+
+            gf2.verify()
+                .await
+                .map_err(|e| VerifierError::MpcError(Box::new(e)))?;
+
+            #[cfg(feature = "tracing")]
+            info!("Finalized all MPC");
+
+            let handshake_summary =
+                HandshakeSummary::new(start_time, server_ephemeral_key, handshake_commitment);
+
+            let session_header = SessionHeader::new(
+                encoder_seed,
+                merkle_root,
+                sent_len,
+                recv_len,
+                handshake_summary,
+            );
+
+            let signature = signer.sign(&session_header.to_bytes());
+
+            #[cfg(feature = "tracing")]
+            info!("Signed session header");
+
+            notarize_channel
+                .send(TlsnMessage::SignedSessionHeader(SignedSessionHeader {
+                    header: session_header.clone(),
+                    signature: signature.into(),
+                }))
+                .await?;
+
+            #[cfg(feature = "tracing")]
+            info!("Sent session header");
+
+            Ok::<_, VerifierError>(session_header)
+        };
+
+        let session_header = futures::select! {
+            res = notarize_fut.fuse() => res?,
+            _ = &mut mux_fut => Err(std::io::Error::from(std::io::ErrorKind::UnexpectedEof))?,
+        };
+
+        let mut mux = mux.into_inner();
+
+        futures::try_join!(mux.close().map_err(VerifierError::from), mux_fut)?;
+
+        Ok(session_header)
+    }
+}
