+++ tlsn/tlsn-verifier/src/tls/state.rs
@@ -8,7 +8,7 @@ use tls_core::key::PublicKey;
 use tls_mpc::MpcTlsFollower;
 
 use crate::{
-    tls::{MuxFuture, OTFuture},
+    tls::future::{MuxFuture, OTFuture},
     Mux,
 };
 
@@ -56,13 +56,56 @@ pub struct Closed {
 
 opaque_debug::implement!(Closed);
 
+/// Notarizing state.
+pub struct Notarize {
+    pub(crate) mux: Mux,
+    pub(crate) mux_fut: MuxFuture,
+
+    pub(crate) vm: DEAPVm<SharedSender, SharedReceiver>,
+    pub(crate) ot_send: SharedSender,
+    pub(crate) ot_recv: SharedReceiver,
+    pub(crate) ot_fut: OTFuture,
+    pub(crate) gf2: ConverterReceiver<Gf2_128, SharedReceiver>,
+
+    pub(crate) encoder_seed: [u8; 32],
+    pub(crate) start_time: u64,
+    pub(crate) server_ephemeral_key: PublicKey,
+    pub(crate) handshake_commitment: Hash,
+    pub(crate) sent_len: usize,
+    pub(crate) recv_len: usize,
+}
+
+impl From<Closed> for Notarize {
+    fn from(value: Closed) -> Self {
+        Self {
+            mux: value.mux,
+            mux_fut: value.mux_fut,
+            vm: value.vm,
+            ot_send: value.ot_send,
+            ot_recv: value.ot_recv,
+            ot_fut: value.ot_fut,
+            gf2: value.gf2,
+            encoder_seed: value.encoder_seed,
+            start_time: value.start_time,
+            server_ephemeral_key: value.server_ephemeral_key,
+            handshake_commitment: value.handshake_commitment,
+            sent_len: value.sent_len,
+            recv_len: value.recv_len,
+        }
+    }
+}
+
+opaque_debug::implement!(Notarize);
+
 impl VerifierState for Initialized {}
 impl VerifierState for Setup {}
 impl VerifierState for Closed {}
+impl VerifierState for Notarize {}
 
 mod sealed {
     pub trait Sealed {}
     impl Sealed for super::Initialized {}
     impl Sealed for super::Setup {}
     impl Sealed for super::Closed {}
+    impl Sealed for super::Notarize {}
 }
